agents/profile-agent/requirements.txt
# Core ADK and A2A dependencies
google-adk[all]>=1.0.0
a2a-sdk>=0.3.0

# Web framework and ASGI server
starlette>=0.46.2
uvicorn[standard]>=0.34.2
fastapi>=0.115.12
sse-starlette>=2.3.5

# HTTP client
httpx>=0.28.1
click>=8.2.0

# OAuth and authentication
authlib>=1.2.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Configuration and data handling
pyyaml>=6.0
pydantic>=2.11.4
python-dotenv>=1.1.0
langchain-google-genai>=2.1.4
langgraph>=0.4.3

# Google Cloud services
google-cloud-secret-manager>=2.16.0
google-cloud-aiplatform>=1.38.0

# Logging and monitoring (optional)
structlog>=23.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # For testing HTTP calls

# Type checking
mypy>=1.5.0
types-PyYAML>=6.0.0